requirements.txt
python-telegram-bot == 21.6
requests == 2.26.0
telebot == 0.0.4
python-dotenv == 0.19.0
emoji==1.6.0
psutil==5.5.1
PyTelegramBotAPi == 4.3.1
